microservices/restapi-service/requirements.txt
adafruit-io==2.8.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.3.2
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.3.0
backoff==2.2.1
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudinary==1.43.0
colorama==0.4.6
commonmark==0.9.1
confluent-kafka==2.9.0
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
dataclasses-json==0.6.7
defusedxml==0.7.1
Django==5.1.7
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
docker==7.1.0
fonttools==4.56.0
frozenlist==1.5.0
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
inference-cli==0.46.5
influxdb3-python==0.12.0
kafka-python==2.1.5
kiwisolver==1.4.8
marshmallow==3.26.1
matplotlib==3.10.1
msgpack==1.1.0
multidict==6.4.3
mypy-extensions==1.0.0
numpy==2.2.4
nvidia-ml-py==12.570.86
opencv-python==4.10.0.84
packaging==24.2
paho-mqtt==2.1.0
pandas==2.2.3
pillow==11.1.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.3
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybase64==1.0.2
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
PyJWT==2.9.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
reactivex==4.0.4
requests==2.32.3
rich==13.0.1
rsa==4.9
scipy==1.15.2
shellingham==1.5.4
six==1.17.0
sqlparse==0.5.3
supervision==0.25.1
tqdm==4.67.1
typer==0.12.5
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
yarl==1.19.0
